pyproject.toml
[project]
name = "hire-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.119.0",
    "google-genai>=1.43.0",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.7",
    "pydantic>=2.12.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sarvamai>=0.1.21",
    "uvicorn>=0.37.0",
]
